moved openwfs wrapper from test to package
